rename stepsForProductAndSurvey -> steps-for-product-and-survey
